feat: refactor for functional test
