+++ src/handlers/delete_client.rs
@@ -1,16 +1,16 @@
+use crate::error::Error::IncludedTenantIdWhenNotNeeded;
 use crate::error::Result;
 use crate::handlers::Response;
 use crate::state::{AppState, State};
 use axum::extract::{Path, State as StateExtractor};
 use std::sync::Arc;
-use crate::error::Error::IncludedTenantIdWhenNotNeeded;
 
 pub async fn handler(
     Path((tenant_id, id)): Path<(String, String)>,
     StateExtractor(state): StateExtractor<Arc<AppState>>,
 ) -> Result<Response> {
     if state.config.default_tenant_id != tenant_id && !state.is_multitenant() {
-        return Err(IncludedTenantIdWhenNotNeeded)
+        return Err(IncludedTenantIdWhenNotNeeded);
     }
 
     state.client_store.delete_client(&tenant_id, &id).await?;
