+++ tests/integration.rs
@@ -1,3 +1,3 @@
 mod env;
 mod providers;
-mod store;
+//mod store; // Comment this out for now
